+++ nodes/Autotask/operations/common/select-columns/index.ts
@@ -1,0 +1,17 @@
+import { selectColumnsOption, addSelectColumnsOption } from './description';
+import {
+	filterEntityBySelectedColumns,
+	filterEntitiesBySelectedColumns,
+	getSelectedColumns
+} from './filter-entity';
+import { prepareIncludeFields, type IPrepareIncludeFieldsOptions } from './include-fields';
+
+export {
+	selectColumnsOption,
+	addSelectColumnsOption,
+	filterEntityBySelectedColumns,
+	filterEntitiesBySelectedColumns,
+	getSelectedColumns,
+	prepareIncludeFields,
+	IPrepareIncludeFieldsOptions
+};